adds example pytest package as PoC
