+++ tests/load/pipeline/test_merge_disposition.py
@@ -36,8 +36,9 @@ from tests.pipeline.utils import (
     load_table_counts,
     select_data,
     load_tables_to_dicts,
-    assert_records_as_set,
 )
+from pytest_dlt.utils import assert_records_as_set
+
 from tests.load.utils import (
     AWS_BUCKET,
     normalize_storage_table_cols,
