+++ tests/pipeline/utils.py
@@ -15,6 +15,7 @@ from dlt.destinations.exceptions import DestinationUndefinedEntity
 from dlt.destinations.dataset.dataset import ReadableDBAPIRelation
 from dlt.common.schema.typing import TTableSchema
 
+from pytest_dlt.utils import assert_records_as_set
 
 PIPELINE_TEST_CASES_PATH = "./tests/pipeline/cases/"
 
@@ -361,27 +362,6 @@ def load_tables_to_dicts(
 
     # done
     return result
-
-
-def assert_records_as_set(actual: List[Dict[str, Any]], expected: List[Dict[str, Any]]) -> None:
-    """Assert that *actual* and *expected* contain the same records irrespective of order.
-
-    Args:
-        actual (List[Dict[str, Any]]): Records retrieved from the destination.
-        expected (List[Dict[str, Any]]): Reference records defined in the test.
-
-    Raises:
-        AssertionError: When the two sets of records differ.
-    """
-
-    def dict_to_tuple(d):
-        # Sort items to ensure consistent ordering
-        return tuple(sorted(d.items()))
-
-    counter1 = Counter(dict_to_tuple(d) for d in actual)
-    counter2 = Counter(dict_to_tuple(d) for d in expected)
-
-    assert counter1 == counter2
 
 
 def assert_only_table_columns(
